requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
gunicorn==20.0.3
Pillow==6.2.1
pytesseract==0.3.0
pytz==2019.3
sqlparse==0.3.0
